Extracted toolbar in a component
