+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/cashpayment/WooPosCashPaymentScreen.kt
@@ -3,12 +3,8 @@ package com.woocommerce.android.ui.woopos.cashpayment
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -21,10 +17,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
@@ -38,6 +32,7 @@ import com.woocommerce.android.ui.woopos.common.composeui.WooPosPreview
 import com.woocommerce.android.ui.woopos.common.composeui.WooPosTheme
 import com.woocommerce.android.ui.woopos.common.composeui.component.WooPosButton
 import com.woocommerce.android.ui.woopos.common.composeui.component.WooPosButtonState
+import com.woocommerce.android.ui.woopos.common.composeui.component.WooPosToolbar
 import com.woocommerce.android.ui.woopos.common.composeui.component.WooPosTypedInputField
 import com.woocommerce.android.ui.woopos.common.composeui.toAdaptivePadding
 import com.woocommerce.android.ui.woopos.root.navigation.WooPosNavigationEvent
@@ -69,7 +64,10 @@ fun WooPosCashPaymentScreen(
     Column(
         modifier = Modifier.fillMaxSize()
     ) {
-        Toolbar(onBackClicked)
+        WooPosToolbar(
+            titleText = stringResource(R.string.woopos_cash_payment_title),
+            onBackClicked = onBackClicked,
+        )
 
         when (state) {
             is WooPosCashPaymentState.Collecting -> {
@@ -183,50 +181,6 @@ private fun Collecting(
                     end.linkTo(parent.end, margin = standardMargin)
                     start.linkTo(parent.start, margin = standardMargin)
                     width = Dimension.fillToConstraints
-                }
-        )
-    }
-}
-
-@Composable
-private fun Toolbar(onBackClicked: () -> Unit) {
-    ConstraintLayout(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 40.dp.toAdaptivePadding())
-            .height(40.dp)
-    ) {
-        val (backButton, title) = createRefs()
-        IconButton(
-            onClick = { onBackClicked() },
-            modifier = Modifier
-                .constrainAs(backButton) {
-                    start.linkTo(parent.start)
-                    top.linkTo(parent.top)
-                    centerVerticallyTo(parent)
-                }
-                .padding(start = 8.dp.toAdaptivePadding())
-        ) {
-            Icon(
-                imageVector = ImageVector.vectorResource(R.drawable.ic_back_24dp),
-                contentDescription = stringResource(R.string.woopos_cart_back_content_description),
-                tint = MaterialTheme.colors.onBackground,
-                modifier = Modifier.size(28.dp)
-            )
-        }
-
-        val iconTitlePadding = 8.dp.toAdaptivePadding()
-        Text(
-            text = stringResource(R.string.woopos_cash_payment_title),
-            style = MaterialTheme.typography.h4,
-            color = MaterialTheme.colors.onBackground,
-            fontWeight = FontWeight.Bold,
-            maxLines = 1,
-            modifier = Modifier
-                .constrainAs(title) {
-                    top.linkTo(backButton.top)
-                    start.linkTo(backButton.end, margin = iconTitlePadding)
-                    centerVerticallyTo(parent)
                 }
         )
     }
