+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/common/composeui/component/WooPosToolbar.kt
@@ -1,0 +1,86 @@
+package com.woocommerce.android.ui.woopos.common.composeui.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.constraintlayout.compose.ConstraintLayout
+import com.woocommerce.android.R
+import com.woocommerce.android.ui.woopos.common.composeui.WooPosPreview
+import com.woocommerce.android.ui.woopos.common.composeui.WooPosTheme
+import com.woocommerce.android.ui.woopos.common.composeui.toAdaptivePadding
+
+@Composable
+fun WooPosToolbar(
+    titleText: String,
+    onBackClicked: () -> Unit
+) {
+    ConstraintLayout(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(top = 40.dp.toAdaptivePadding())
+            .height(40.dp)
+    ) {
+        val (backButton, title) = createRefs()
+        IconButton(
+            onClick = { onBackClicked() },
+            modifier = Modifier
+                .constrainAs(backButton) {
+                    start.linkTo(parent.start)
+                    top.linkTo(parent.top)
+                    centerVerticallyTo(parent)
+                }
+                .padding(start = 8.dp.toAdaptivePadding())
+        ) {
+            Icon(
+                imageVector = ImageVector.vectorResource(R.drawable.ic_back_24dp),
+                contentDescription = stringResource(R.string.woopos_toolbar_icon_content_description),
+                tint = MaterialTheme.colors.onBackground,
+                modifier = Modifier.size(28.dp)
+            )
+        }
+
+        val iconTitlePadding = 8.dp.toAdaptivePadding()
+        Text(
+            text = titleText,
+            style = MaterialTheme.typography.h4,
+            color = MaterialTheme.colors.onBackground,
+            fontWeight = FontWeight.Bold,
+            maxLines = 1,
+            modifier = Modifier
+                .constrainAs(title) {
+                    top.linkTo(backButton.top)
+                    start.linkTo(backButton.end, margin = iconTitlePadding)
+                    centerVerticallyTo(parent)
+                }
+        )
+    }
+}
+
+@WooPosPreview
+@Composable
+fun WooPosToolbarPreview() {
+    WooPosTheme {
+        Column {
+            WooPosToolbar(
+                titleText = "Title",
+                onBackClicked = { }
+            )
+
+            Spacer(modifier = Modifier.weight(1f))
+        }
+    }
+}
